move: globalStyle을 style 폴더로 이동
