+++ src/App.tsx
@@ -1,8 +1,8 @@
 import { RouterProvider } from "react-router-dom";
 
 import QueryClient from "./core/queryClient";
-import GlobalStyle from "./globalStyle";
 import router from "./routes/route";
+import GlobalStyle from "./styles/globalStyle";
 
 export const queryClient = new QueryClient();
